+++ src/routes/Layout.svelte
@@ -36,8 +36,8 @@
         <li>We will <b>no longer collect any data</b> for our platform or studies</li>
         <li>You will be <b>removed from any studies</b> you're currently participating in</li>
         <li>We will <b>delete any demographic information</b> you've shared with us</li>
-        <li>Researchers will <b>not receive any additional data</b> from you, but will <b>retain access to the data</b> you've already shared with them.</li>
-        <li>We will <b>uninstall & remove the Rally add-on</b>, and any associated study add-ons that you may have installed.</li>
+        <li>If you are currently enrolled in an open study, we will <b>delete all data</b> we've collected through that study to date. However, researchers will still retain access to data you've contributed to <b>completed studies</b>.</li>
+        <li>We will <b>uninstall and remove the Rally add-on</b>, and any associated study add-ons that you may have installed.</li>
       </ul> 
     </div>
     <img style="width: 270px; padding-top: 40px; transform: translateX(-24px);" src="img/before-you-go.png" alt="person walking through exit door" />