+++ stories/studies/StudyJoinFlow.svelte
@@ -11,7 +11,7 @@
   // add one nicer study with all the information.
   const nicerStudy = {
     addonId: 'nicer-study-example@mozilla.org',
-    name: "Rally Core Study",
+    name: "Test STudy",
     authors: {
       name: "Rally Team"
     },
@@ -21,8 +21,7 @@
     endDate: new Date('2021-07-03'),
     tags: ['ad trackers', 'misinformation', 'profiling'],
     privacyPolicy: {spec: 'https://example.com'},
-    description: `This study is a a simple scaffold. Researchers can add any description they prefer here.
-    Perpetual sincerity out suspected necessary one but provision satisfied. Respect nothing use set waiting pursuit nay you looking. If on prevailed concluded ye abilities. Address say you new but minuter greater. Do denied agreed in innate. Can and middletons thoroughly themselves him. Tolerably sportsmen belonging in september no am immediate newspaper. Theirs expect dinner it pretty indeed having no of. Principle september she conveying did eat may extensive. `,
+    description: `This is a mock study. All of the information in this study card is strictly for this example.`,
     dataCollectionDetails: ['page views', 'time and date of joining study', 'etc.'],
 
   }
@@ -35,11 +34,11 @@
     });
 
   function toggleStudyJoinStatus(studyID, joined) {
-    // set study.studyInstalled if studyID = study.study_id.
+    // set study.studyJoined if studyID = study.study_id.
     studies = [...studies.map(study => {
       const nextStudy = {...study};
       if (nextStudy.addonId === studyID) {
-        nextStudy.studyInstalled = joined;
+        nextStudy.studyJoined = joined;
       }
       return nextStudy;
     })];